refactor(homepage): extract out footer link as a new component
